common update and delete method

+++ controller/userController.js
@@ -1,5 +1,6 @@
 import HttpError from "../middleware/HttpError.js";
 import User from "../model/User.js";
+import cloudinary from "../config/cloudinary.js"
 
 const add = async (req, res, next) => {
   try {
@@ -106,7 +107,9 @@ const allUser = async (req, res, next) => {
 
 const update = async (req, res, next) => {
   try {
-    const user = req.user;
+    const targetedUser = req.params.id || req.user._id;
+
+    const user = await User.findById(targetedUser)
 
     if (!user) {
       return next(new HttpError("user not found", 404));
@@ -114,17 +117,35 @@ const update = async (req, res, next) => {
 
     const updates = Object.keys(req.body);
 
-    let allowedFields = ["name", "password", "phone"];
+    let allowedFields = ["name", "password", "phone", "profilePic"];
 
 
-    if (req.body.role === "admin" || req.body.role === "super_admin") {
-      const allowedFields = [...allowedFields, role, profilePic]
+    if (req.user.role === "admin" || req.user.role === "super_admin") {
+      allowedFields = [...allowedFields, "role", "isVerified"]
     }
 
     const isValid = updates.every((field) => allowedFields.includes(field));
 
     if (!isValid) {
       return next(new HttpError("only allowed field can be updated", 400));
+    }
+
+    if (
+      !req.user.role === "admin" &&
+      !req.user.role === "super_admin" &&
+      !req.user._id.toString() !== user, _id.toString()
+    ) {
+      return next(new HttpError("unauthorized access", 401))
+    }
+
+    if (req.file) {
+      if (user.cloudinaryId) {
+        await cloudinary.uploader.destroy(user.cloudinaryId)
+      }
+
+      user.profilePic = req.file.path;
+
+      user.cloudinaryId = req.file.fileName;
     }
 
     updates.forEach((update) => (user[update] = req.body[update]));
@@ -141,9 +162,27 @@ const update = async (req, res, next) => {
 
 const deleteUser = async (req, res, next) => {
   try {
-    const user = req.user;
+    const targetedUser = req.params.id || req.user._id;
 
-    await User.deleteOne(user);
+    const user = await User.findById(targetedUser)
+
+    if (!user) {
+      return next(new HttpError("user not found", 404));
+    }
+
+    if (
+      !req.user.role === "admin" &&
+      !req.user.role === "super_admin" &&
+      !req.user._id.toString() !== user._id.toString()
+    ) {
+      return next(new HttpError("unauthorized access", 401));
+    }
+
+    await User.deleteOne(user)
+
+    if (user.cloudinaryId) {
+      await cloudinary.uploader.destroy(user.cloudinaryId)
+    }
 
     res
       .status(200)
